applications/backend/app/requirements.txt
flask==3.0.3
mysql-connector-python==9.0.0
pytest==8.4.1
